feat: 오늘의 학습 단어 / 학습 이벤트 기록 엔드포인트 + 프론트 학습 흐름 정비
- WordController: GET /api/words/daily (사용자별 오늘의 추천 20단어), POST /api/words/{wordId}/view (PBI-12 STUDY_WORDS 미션 트리거) 추가
- WrongNoteProjection/Response: MySQL COUNT(*) BIGINT 반환에 맞춰 wrongCount 타입을 Integer/int → Long/long 으로 교정 (Hibernate 캐스팅 실패 수정)
- 프론트 wordsApi/wrongNotesApi: 백엔드 변경에 맞춰 userId 쿼리 파라미터 전달
- NavMenu: '단어' → '단어 학습'(/flashcard) + '단어 목록'(/words) 로 분리
- useSpeech: 발음 속도(rate) 파라미터 추가, 연타 시 끊김 방지 cancel() 주석화
- 대시보드/플래시카드/단어 목록/오답노트 UI 정비 (TodayWordsCard, WordCard, WordFilterBar, PronounceButton, WrongNoteList, DashboardPage, FlashcardPage, WordListPage, WrongNotePage)

+++ backend/src/main/java/com/uhheung/voca/wrongnote/dto/WrongNoteProjection.java
@@ -6,6 +6,7 @@ public interface WrongNoteProjection {
     Long getWordId();
     String getEnglish();
     String getKorean();
-    Integer getWrongCount();
+    // MySQL COUNT(*) 는 BIGINT 를 반환하므로 Long 으로 받는다. Integer 로 받으면 Hibernate 캐스팅 실패.
+    Long getWrongCount();
     LocalDateTime getLastWrongAt();
 }
